AI/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.30.1
pymongo==4.8.0
python-dotenv==1.0.1
requests==2.32.3
PyPDF2==3.0.1
sentence-transformers==2.7.0
langdetect==1.0.9
deep-translator==1.11.4
httpx==0.27.2
numpy==2.1.3
torch --index-url https://download.pytorch.org/whl/cpu